Modified the movement and added arrow key teleportation.
Modified the movement so that it detects that either the Left or Right shift key is held down during movement in order to activate a boost. I also made it so the arrow keys would teleport the player in the direction of the arrow key, based on the sprites current orientation. Also moved the random teleport functions and variables form the SpritMover into the controller and pawn scripts.