Modified a lot of stuff. GUI and Health component changes.
Modified a lot of stuff. Added UI for scoring, a timer, and counting the number of enemies in the game. I had to remake the Health component and make two new classes for player and enemy health. Made a no-rotation script so that the health bar added to the meteorites won't rotate with the meteor. Made a png of a stopwatch to replace the png for the timer.